+++ source/performance/F0.Generators.Benchmarks/CodeAnalysis/EnumInfoGeneratorBenchmarks.cs
@@ -13,13 +13,14 @@ public class EnumInfoGeneratorBenchmarks
 		string code =
 @"#nullable enable
 using System;
+using System.Reflection;
 using F0.Generated;
 
 public sealed class Class
 {
-	public void Method(StringSplitOptions options, DateTimeKind kind)
+	public void Method(ResourceLocation flags, DateTimeKind kind)
 	{
-		_ = EnumInfo.GetName(options);
+		_ = EnumInfo.GetName(flags);
 		_ = EnumInfo.GetName(StringComparison.CurrentCultureIgnoreCase);
 
 		_ = EnumInfo.GetName(kind);
@@ -76,9 +77,15 @@ namespace F0.Generated
 			throw new global::F0.Generated.SourceGenerationException($""Cannot use the unspecialized method, which serves as a placeholder for the generator. Enum-Type {{value?.GetType().ToString() ?? ""<null>""}} must be concrete to generate the allocation-free variant of {nameof(Enum)}.{nameof(Enum.ToString)}()."");
 		}}
 
-		public static string GetName(global::System.StringSplitOptions value)
+		public static string GetName(global::System.Reflection.ResourceLocation value)
 		{{
-			throw new global::F0.Generated.SourceGenerationException(""Flags are not yet supported: see https://github.com/Flash0ver/F0.Generators/issues/1"");
+			return value switch
+			{{
+				global::System.Reflection.ResourceLocation.ContainedInAnotherAssembly => nameof(global::System.Reflection.ResourceLocation.ContainedInAnotherAssembly),
+				global::System.Reflection.ResourceLocation.ContainedInManifestFile => nameof(global::System.Reflection.ResourceLocation.ContainedInManifestFile),
+				global::System.Reflection.ResourceLocation.Embedded => nameof(global::System.Reflection.ResourceLocation.Embedded),
+				_ => throw new global::System.ComponentModel.InvalidEnumArgumentException(nameof(value), (int)value, typeof(global::System.Reflection.ResourceLocation)),
+			}};
 		}}
 
 		public static string GetName(global::System.StringComparison value)