+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 CHANGELOG
 
 ## vNext
+- Changed `EnumInfoGenerator`, additionally generating strongly typed overloads of `EnumInfo.GetName(System.Enum)` for enumeration types with the `System.FlagsAttribute` applied, equivalent to non-Flags.
 
 ## v0.2.1 (2021-08-08)
 - Fixed `EnumInfoGenerator`, no longer throwing a `System.OverflowException` rather than the intended `System.ComponentModel.InvalidEnumArgumentException`, in case of the default overflow/underflow checking context is set to _checked_ when invoking `EnumInfo.GetName(System.Enum)` with a value which neither has an associated enum member in that enumeration type nor is within the range of `int`.